[Refactor/#203] RewardPoint 도메인 모델 위치를 reward 패키지로 이
