+++ feature/my/src/main/java/com/ilsangtech/ilsang/feature/my/screens/mytab/component/MyObtainedPointContent.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.ilsangtech.ilsang.core.model.RewardPoint
+import com.ilsangtech.ilsang.core.model.reward.RewardPoint
 import com.ilsangtech.ilsang.core.ui.season.model.SeasonUiModel
 import com.ilsangtech.ilsang.core.ui.user.UserObtainedPointCard
 import com.ilsangtech.ilsang.core.ui.user.model.UserObtainedPointUiModel
