+++ feature/submit/src/main/java/com/ilsangtech/ilsang/feature/submit/model/SubmitResultUiState.kt
@@ -1,6 +1,6 @@
 package com.ilsangtech.ilsang.feature.submit.model
 
-import com.ilsangtech.ilsang.core.model.RewardPoint
+import com.ilsangtech.ilsang.core.model.reward.RewardPoint
 
 sealed interface SubmitResultUiState {
     data object NotSubmitted : SubmitResultUiState
